:recycle: Refactored services & added validation control
